Cloud Scheduling Techniques.
Problem Statement
To design effective task scheduling algorithms that have the ability to solve NP-Hard Problem in such a way that whenever a new task is ready to be executed, we need to assign it to a server so as to minimize the completion time of all the tasks. The objective throws a light on comparative analysis of different task scheduling algorithms on the basis of their makespan.